requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.11
faiss-cpu>=1.7.4
sentence-transformers>=2.3.1
ollama>=0.1.6
pypdf>=3.17.4
python-multipart>=0.0.9
python-magic>=0.4.27  # For file type detection
llama_cpp